Change equipment inventory screen UI
Equipment inventory screen needs to be updated in order to support cold chain equipment (CCE). UI needs to be improved, but some work needs to be done in order to get facilities, programs and equipment types to work together in the screen. The tricky part is that for "my facility" the facility is retrieved first, then the program, but for supervised facilities, the program is retrieved first.

Changes:

* Move my facility text box to just text under "My facility"
* Remove facility dropdown, as screen should show equipment for all supervised facilities
* Add equipment type dropdown