Adjust get inventory service call
Getting inventory from the backend needs to be changed in order to get by selection (my facility or supervised facilities), program and equipment type. It should also not need facility ids from the frontend, as that can be determined by the logged in user id and the selection. This ensures the server only gets the inventory of facilities the user has access to.

The lower service layer does the logic of getting the list of valid facilities and getting its corresponding inventory. Also, as a result, getting facilities is not necessary on the frontend and can be removed.

Also fix the test cases.