Add isColdChain element to equipment type
This adds the isColdChain boolean to equipment types. This is in order to check for the value in showing different tables in the inventory list screen based on selecting an equipment type that is cold chain or not. This will also affect screens in the equipment and equipment type management screen, which should be implemented in a future commit.