Add update status dropdown to inventory list screen
User needs an easier way to update status, directly from the inventory list screen instead of having to go into the edit screen to do that. Create dropdowns with possible status values, and update the server when they change.

Also need to create feedback when status is changed, message is shown when change is successful. Message stays on for 3 seconds and then by animation, gradually disappears.