Refactor attributes out of equipment inventory model
Certain attributes stored in inventory should be removed or refactored out of equipment inventory model. This needs to be done in the model, its tests and other models related to it (requisition, vaccine report, equipment line items). Certain views also need to be modified to reference attributes in different tables.