Add additional statuses to inventory screens
Need to create additional statuses for CCE, Functional, Functional But Not Installed, Not Functional. Add them in the database. Also need to create new field in status table, category, to differentiate between lab equipment statuses and CCE statuses. Change UI to only show certain statuses, depending on equipment type.

Also need to create reason for not functional statuses. Add them in the database and the service layer. From this, need to create new field in status table, called isBad, to determine if a status is "bad" in order for the UI to check if a "bad" status has been selected.

Adjust test cases to reflect new code. Enable test cases for EquipmentInventoryMapper.