Add modal for not functional status in inventory list screen
When selecting a status as "Not Functional", show a modal that gives further questions, such as reason for non-function, and an option to make not active and decommissioned date if obsolete. This is to avoid having to go to the edit screen to make these changes. Also add test cases to test additional functionality.