Add pagination to equipment inventory list screen
Pagination is necessary on inventory list screen, to not retrieve all results if there are many. Pagination is set up in the controller and actually done in the SQL. In order to help do pagination, add a get inventory count method to backend.

On frontend, to get pagination to work properly, add ui-bootstrap-pagination in th
e index.html, add it in module definition in module.js, add page to arguments in t
he service call and set scope variables on return, add pagination directive in the list.html, and add a watch to currentPage scope variable to load inventory using the new value of currentPage.

Fix test cases to include pagination. Also add test cases for getting inventory count and a missing test case in the mapper.

Also add helper function to get facility id string from facility ids in repository, and helper function to get facility ids from user and type in service.