Refactor code field from equipments table
The add inventory screen does not know what to fill in for the code field, and that information is more appropriately stored in the equipment_types table, so it makes sense to remove it from the equipment list.

Add a migration script, remove from the mapper, fix test cases (and remove two unnecessary test cases).