Add 'save new equipment' functionality in add inventory screen
Non cold chain equipment is managed at a lower level, and those people do not have access to the equipment list page to add new equipment. Therefore, they need to be able to do that through the add inventory screen.

Basically, the user enters manufacturer, model (and energy type) information. When saving, the backend checks if the info entered matches an entry in the database. If not, insert a row for equipment. Otherwise, update the row. Then,save the inventory item (possibly using the new equipmentId if equipment was added).

Change frontend to replace equipment dropdown with equipment type readonly textbox. Also set equipment name right before save. On backend, for non-CCE, insert or update equipment.