Add stabilizer field to equipment inventory
Stabilizer field was not present as a data field for equipment inventory. Add it as a column to the equipment inventory table, include it in the equipment inventory DTO, and adjust the mapper code to include it in the insert and update.

Change the front-end from a dropdown with yes/no to a checkbox and include an entry in the messages file.