Fix update equipment on inventory save
For non-CCE, when saving inventory, sometimes equipment also needs to be saved. In cases where the equipment already exists in the db, the equipmentId needs to be set in the inventory object to avoid a SQL exception.

Also fix a save problem when user does not specify facility (for adding to a supervised facility) by requiring facility to be set on the client side.