VIMS-211 Disallow remove equipment if inventory exists
Equipment entries are not removable if inventory exists for them. Currently, the behavior is to allow them, but when a user attempts to do so, an exception is thrown and logged, and nothing happens on the user's screen. To avoid this, hide the remove link from the equipment list.

To do so, add an inventory count property to the Equipment object, and populate it in the mapper code. Then, check for an inventory count in the frontend greater than zero and hide accordingly.