On remove equipment with existing inventory, return error
In case there is a remove equipment call to the server, do a check to see if equipment is removable and throw error if not.

Add isRemovable method to Equipment class and have repository code use it to check for inventory existence. Add inventorycount to other mapper methods so that inventoryCount field is filled in for Equipment object.

Add frontend code to show error message in case remove fails, showing error message from the server.