Set notFunctionalStatusId to null when status id is "good"
When status id is set to a "good" id, not functional status id should be invalid. Set it to null when status id is set. This is more logical and will help with reporting. Also, create a new mapper function to get statuses by id. Set not functional status id to null on the frontend, just for UI.