Ameliorate swagger-related issue
The version of Swagger-UI that we’re using erroneously shows all URIs as deprecated. This check-in represents the path of least resistance to addressing this. (Upgrading Swagger UI and perhaps Swagger-Core/Swagger-Annotations can be revisited in the future. It involves more time, however, along with unnecessary side affects.) This check-in also tweaks the Swagger page such that it prompts users to login to eLMIS if need be.