VIMS-332 Update adjustment reason data
Need to update list of adjustment reasons, but in a way that is not just for VIMS. Create upload support for adjustment reasons. Do this by creating handlers and adding entries in the XML configuration and annotating the adjustment reason classes.

Also need a way to support which adjustment reasons are for which program, so create a mapping table between reasons and programs (stock_adjustment_reasons_programs table). Create upload support for this as well (handlers, XML config, annotations).

Create backend code to support these uploads (service, repo, mapper).