Add isdefault to stock adjustment reasons
New property isdefault is necessary to get a list of "default" adjustment reasons, for any programs not in the mapping table (stock_adjustment_reasons_programs). This is to keep the list of existing adjustment reasons for the RnR process.

Create schema migration script and add code to repo/mapper to only get default reasons, not all.