Add test case for parent calculation of FPP ISA upload
Add additional test case that tests that parent calculations are correct for the facility program product ISA upload handler. Change some handler code to remove redundant object.

Add setOverriddenIsa call in save method in order to get the test case to work. The test case needs to check the wastage factor value for the calculated parent facilities, but they do not show up in the fpp object because the verify method only checks the fields of the class, not any of the fields of the superclass. Since overriddenIsa is in the class and it does not negatively impact to set it and it helps to make the test case work, it is added.