Compute wastage factor based on weighted average
Computer based on weighted average, not straight average. Meaning DVS, RVS and CVS will be calculated by looking at facility wastage factors weighted by population.

A service in the demographics module is used. This code is refactored out of the StockRequirementsService into its own class, PopulationService, since it is generic enough, and the getPopulation() code is moved to the new class.

Handler code uses getPopulation() in demographics module to get facility population. However, this does not initially work because then the core module would need to depend on the demographics module, which would create a circular reference since the demographics module already depends on the core module. Therefore, the handler class, its test class and the DTO is moved to the vaccine module/package.

DTO adds a Population Source value using its name in order to determine which population to use during upload. Name is used instead of ID, to be more user-friendly in the upload file. However, this requires a new method in the EstimateCategoryRepository and mapper, getByName, to get an ID by name. This ID is then passed to the getPopulation() method.