Make changes based on code review of pull request
Population source ID is set to null, rather than a default, and rather than throwing an exception. Default should not be used, but an exception should not be thrown because a population source does not have to be specified in order for a population to be retrieved, since the facility's catchment population can be used.

Also, for VIMS, the definition of an SDP is not to check the SDP flag, but to see if the facility type's code is "heac" or "disp"(!). Change the code to reflect that.