Remove and change code to get build to work
After merging in stock management code out of its own repository and into open-lmis, there are build errors. This is because some code was changed in the stock-management repository that broke builds. That code is reverted and put into another branch, however, calls to that code is still in the main open-lmis repo. These changes call the original (reverted) code.