Update PostgreSQL to 9.2 in gradle build scripts
OpenLMIS 2.0 requires PostgreSQL to be at 9.2+ in order to build properly. This should be reflected in the build files.

One snag is that once 9.2+ jars are used, it seems in order to use JDBC, you must have a trailing slash in the URL, or else a "no suitable driver found" error is given.