OLMIS-55 Update gradle build scripts to flyway 3.2
Update gradle build's flyway to 3.2. Use the gradle-flyway plugin provided by Flyway.

It appears that flyway-ant had a dependency to log4j and slf4j, so once it was removed, those dependencies needed to be added somewhere else, so they are added in core and openlmis-web.

Fix the RightMapper integration test to decrement the number of rights by 1.