OLMIS-85 Fix distribution bug
Distribution cannot be initiated because of a Jackson error. Basically the Distribution class has a status field which is a DistributionStatus. This has not been able to be serialized properly by Jackson 2.x and Spring for some reason. A custom serializer StatusSerializer (and deserializer), is used to do the serialization for Jackson 1.9.9 (and the corresponding older Spring version), but when Spring was updated, Jackson was as well, and somehow the custom serializer does not work with the new Jackson.

The fix is to not use a custom serializer and deserializer class, but use JsonFormat, which is new in Jackson 2, saying serialize it as an object. No corresponding deserialization is implemented, as there is no use for it currently. The custom classes are no longer used, so they areremoved.