Create devadmin user for testseed task
Create a user for developer testing (devadmin), giving it all permissions in the system. This is to test toggle functionality and for other future developer testing purposes.

This user should not be used for full R&R testing, as it only has a home facility assigned to one program, and no supervisory nodes assigned.