Fix OutOfMemoryError when running gradle build jetty server
There is an old OutOfMemoryError issue when running jetty in the gradle build. This is because of not enough memory for Jetty. The workaround has been to have the builder export JAVA_OPTS and JAVA_TOOL_OPTIONS to up the memory. We can fix this by setting a gradle property to up the memory when running gradle.