Replace seed scripts with new seed scripts
Seed scripts need to be changed, as some migration scripts would do data inserts, which are more appropriate in the seed task. Original scripts can not be preserved, as new scripts are generated by pg_dump and a challenge to coordinate the original scripts with the new dump generated scripts.

Seed script has two generated scripts, one a baseline seed that is more of a true seeding script, and an extra seed that adds data that can be added in other ways (upload, UI), so not a true seed. This extra seed is necessary to be added because the integration tests depend on it. Eventually the extra seed should be removed from the seed task.