Add additional toggle off features
These additional toggle off features require some additional rights to be added to the system, mainly for hiding certain UI elements.

For the hasPermission check to work in a part of the UI, the hasPermission function from the AuthorizationService has to be put into the controller's scope for the HTML page. Which also means the AuthorizationService needs to be a dependency on that controller.

Also need to add new messages to the properties file.