Add and rearrange gradle tasks to separate seed process
Need to separate base seed and extra seed tasks, so that each can be called independently of each other. Create a separate extraSeed task for this, calling the extra_seed.sql script. Create a globalSeed task to call both baseSeed and extraSeed.

Rename baseline_seed to true_seed, so it is not confused as the only script that is run by the baseSeed task.