OLMIS-137 Fix error when listing equipment by program
When program is selected in the equipment list page, an error is logged in the log file. This is because the equipment list is attempted to be retrieved, even though the equipment type is not specified. Do a check that equipment type is set to something before making the call to the equipment list.