OLMIS-136 Fix blank PQS designation field in edit
PQS designation field is blank in the edit page because when using an object for ngModel with ngOptions, setting a preselected value will not work, because ngModel checks by reference, not by value. The solution here is to use ngOptions "track by" to track by the object's id property, rather than the object itself.