Use environment variable for configuration
OpenLMIS originally allowed the location of its app.properties file to be specified via an environment variable called environmentName. VillageReach's deployment scripts rely on this flexibility. For some reason, though, it was removed. This commit restores this functionality, taking care to leave reference to the database-config.properties file which isn’t part of our known configuration file set, but which may nevertheless be used on existing deployments.