OLMIS-135 Fix visual bug in Manage POD
In Manage POD page, when selecting program and facility, the list does not show properly because of a ng-grid visual glitch. Not sure what is causing it, but resetting scope.orders to an empty array before populating seems to fix it.

Also fixed another bug where deselecting the checkbox empties the order list, rather than showing the list filtered by facility.