OLMIS-160: Revamp Offline Functionality
At some point, seemingly incomplete and breaking changes were made to OpenLMIS' offline-related functionality. For example, all of the menu-items were removed from the offline version of the menubar, leaving users with a blank swath of green which obscures most menu items in the non-offline version of the menubar. Functionality related to the "Home" link seems similarly to have been changed but left incomplete. This commit thus:

1. Reintroduces OpenLMIS' original implementation of offline-functionality.

2. Updates the manifest such that it caches the "LOGIN_SUCCESS_DEFAULT_LANDING_PAGE" configuration.

3. Updates the manifest such that it caches additional fonts. The UI isn't consistent in terms of how it requests its fonts. Until this is improved, it's necessary to cache several redundant ones.