OLMIS-161: Make "Go Online" Button More Robust
As the following steps aim to show, the "Go Online" button is currently a little brittle:

1. Clear your browser's cache and then log into OpenLMIS.
2. Go offline, and then request a page in OpenLMIS that's unrelated to distributions. (The goal is to request a page which isn't listed in the manifest as being available offline.)
3. Observe that you're properly shown the offline menu.
4. Go back online (perhaps via Chrome's dev tools if that's how you went offline at the beginning of step 2.) Then, now that you have connectivity, click "Go Online" within OpenLMIS.
5. Observe that nothing seems to happen. This is because the "Go Online" button assumes you're currently on a page intended for offline use.

NOTE: None of the above applies if the page you requested in step 2 is already cached on your system. In that case, the browser serves the cached page, which then behaves oddly because it can't retrieve the data it's reliant on from the backend.