OLMIS-147 Fix save after password reset unverifying user
Reset password code verifies unverified user; however, doing a save in the edit user screen updates the user again, and for a previously unverified user, that "Is verified" field is "No", so user is inadvertently unverified. This is because of isMobileUser code saving the verify field in the database, which was previously ignored. Re-exclude the verify field from the update method in the mapper and create a separate verify mapper method. Use that method for the isMobileUser code to automatically verify the user.